Always write tmp files to cache when not on primary external storage
If we process files from other storage than primary, and that settings
are to not write tmp files to cache, there was a "operation not
permitted" when creating the temporary file holding the thumnail for exiv2
or libexif. This is because we use a "File" to make operations with
these native libraries. And the tmp files cannot be written with File
classes into 2ndary external storage.

In this case, we force to use the externalCacheDir although the user
doesn't want