Fix incorrect exif attribute PIXEL_X_DIMENSION & PIXEL_Y_DIMENSION
When there was no initial EXIF data in a picture, the PIXEL_X_DIMENSION & PIXEL_Y_DIMENSION
attributes were either missing or set to 0.

In that case, Windows wasn't able to show the thumbnails correctlty in the explorer or
import tools.

Setting and adding these fields correctly fixes the issue.

Fixes: https://github.com/tenzap/exif-thumbnail-adder/issues/24
Fixes: https://github.com/tenzap/exif-thumbnail-adder/issues/25