add setting to fix pictures having thumbnail but some tags missing
It can happen that some pictures already have a thumbnail but don't show up
in Windows when importing picutres. This is because some of the mandatory
tags are missing in the EXIF metadata.

So add an option to permit update of the files that already have a thumbnail
but don't have all mandatory fields.

Enabling the option will replace the thumbnail that was already in the file